requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-nest==4.2.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
APScheduler==3.10.4
flask-swagger-ui==5.21.0